% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_basicos_legendas.R
\name{Dados_basicos_legenda}
\alias{Dados_basicos_legenda}
\title{Dados_basicos_legenda(x, y)}
\usage{
Dados_basicos_legenda(Insira_Link_do_Video_aqui, language = "en-US")
}
\arguments{
\item{x}{é string com link para video do youtube.}

\item{y}{é um string indicando qual o idioma da legenda, o valor
default é "en-US". Para entender quais os possiveis strings de 'y', ver o
parametro language da função get_caption() do pacote "youtubecaption"}
}
\description{
A funçaõ retorna dados básicos gerais sobre o texto das legendas.
}
