% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_basicos_legendas.R
\name{Dados_basicos_legenda_vetor}
\alias{Dados_basicos_legenda_vetor}
\title{Dados_basicos_legenda_vetor(x,y)}
\usage{
Dados_basicos_legenda_vetor(Insira_vetor_de_Links_aqui)
}
\arguments{
\item{x}{é vetor de strings com links para video do youtube.}

\item{y}{é um string indicando qual o idioma da legenda, o valor default é
"en-US". Para entender quais os possiveis strings de 'y', ver o parametro
language da função get_caption() do pacote "youtubecaption".
ATENÇÃO: não foi construída uma forma de indicar a lingua de cada legenda.
Portanto, é necessário que todos os links de videos adicionados tenham a
legenda desejada no mesmo idioma.}
}
\description{
A funçaõ retorna  um data.frame com dados básicos gerais sobre o texto
das legendas.
}
