% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_caption.R
\name{legenda_sent_completa}
\alias{legenda_sent_completa}
\title{legenda_sent_completa(x, y)}
\usage{
legenda_sent_completa(Insira_Link_do_Video_aqui, languag = "en-US")
}
\arguments{
\item{x}{é string com link para legenda;}

\item{y}{é string indicando qual o idioma da legenda,
o valor default é "en-US". Para entender quais os possiveis strings de 'y',
ver o parametro language da função get_caption() do pacote "youtubecaption"}
}
\description{
A função retorna um DF com a legenda organizada em sentenças completas a cada linha.
}
