% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listasTematicas_davies&gardner.R
\name{tematicasDGvariantes}
\alias{tematicasDGvariantes}
\title{CRIANDO UMA FUNCAO PARA RETORNAR A BASE DE DADOS COM TODAS AS PALAVRAS TEMATICAS (com palavra base e variantes)
tematicasDGvariantes()
função sem parametro. retorna DF com as listas tematicas de Davies & Gardner.
No entanto, elas está expandida, pois contem (1) uma coluna com os lemmas
(lista original com valores repetidos para os tokens de mesmo lemma);
(2) coluna de tokens (as tokens variantes dos lemmas foram obtidas
com a base hash_lemmas do pacote lexicon)
(3) coluna 'group' que agrupa as palavras por grupos temáticos
(são eles: months, weekdays, seasons, intervals, dayparts, holidays, time_general,
emotions_negative, emotions_neutral, emotions_positive, color, clothing,
clothing_general, animals, bodyparts, family, food, food_general, material,
 nationalities, professions, professions_general, government, military,
 sports_general, sports_participants, sports, transportation_general,
 transportation, weather_general, weather_nouns)}
\usage{
tematicasDGvariantes()
}
\description{
CRIANDO UMA FUNCAO PARA RETORNAR A BASE DE DADOS COM TODAS AS PALAVRAS TEMATICAS (com palavra base e variantes)
tematicasDGvariantes()
função sem parametro. retorna DF com as listas tematicas de Davies & Gardner.
No entanto, elas está expandida, pois contem (1) uma coluna com os lemmas
(lista original com valores repetidos para os tokens de mesmo lemma);
(2) coluna de tokens (as tokens variantes dos lemmas foram obtidas
com a base hash_lemmas do pacote lexicon)
(3) coluna 'group' que agrupa as palavras por grupos temáticos
(são eles: months, weekdays, seasons, intervals, dayparts, holidays, time_general,
emotions_negative, emotions_neutral, emotions_positive, color, clothing,
clothing_general, animals, bodyparts, family, food, food_general, material,
 nationalities, professions, professions_general, government, military,
 sports_general, sports_participants, sports, transportation_general,
 transportation, weather_general, weather_nouns)
}
